pyproject.toml
[project]
name = "max"
version = "0.1.0"
description = "Max Bot with GigaChat integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff",
    "python-dotenv",
    "requests>=2.32.5",
    "gigachat>=0.1.0",
    "lightrag-hku>=1.4.9.8",
    "nano-vectordb",
    "networkx>=3.0",
    "numpy>=1.24.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I"]
fix = true


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
